# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hakim Medical History Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Patient {
  id: number;
  name: string;
  age?: number | null;
  gender?: string | null;
  occupation?: string | null;
  weight?: string | null;
  height?: string | null;
  maritalStatus?: string | null;
  email?: string | null;
  phone?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type PatientProfileCurrentMedicationsItem = { [key: string]: unknown };

export type PatientProfileAllergiesItem = { [key: string]: unknown };

export type PatientProfileFamilyHistory = { [key: string]: unknown };

export interface PatientProfile {
  id: number;
  patientId: number;
  chronicConditions?: string[];
  surgicalHistory?: string[];
  currentMedications?: PatientProfileCurrentMedicationsItem[];
  allergies?: PatientProfileAllergiesItem[];
  familyHistory?: PatientProfileFamilyHistory;
  smokingStatus?: string | null;
  alcoholUse?: string | null;
  createdAt: string;
}

export type CaseRosSymptoms = { [key: string]: unknown } | null;

export type CaseChiefComplaintsItem = { [key: string]: unknown };

export interface Case {
  id: number;
  patientId: number;
  status: string;
  rosSymptoms?: CaseRosSymptoms;
  chiefComplaints?: CaseChiefComplaintsItem[] | null;
  createdAt: string;
  updatedAt: string;
}

export interface PatientFull {
  patient: Patient;
  profile?: PatientProfile | null;
  cases: Case[];
}

export type InterviewMessageRole =
  (typeof InterviewMessageRole)[keyof typeof InterviewMessageRole];

export const InterviewMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface InterviewMessage {
  role: InterviewMessageRole;
  content: string;
  timestamp?: string;
}

export interface InterviewSession {
  id: number;
  caseId: number;
  messages: InterviewMessage[];
  isComplete: string;
  createdAt: string;
  updatedAt: string;
}

export type ReportPatientInfo = { [key: string]: unknown };

export interface Report {
  id: number;
  caseId: number;
  patientInfo: ReportPatientInfo;
  chiefComplaint?: string | null;
  hpi?: string | null;
  ros?: string | null;
  pmh?: string | null;
  drugHistory?: string | null;
  allergyHistory?: string | null;
  familyHistory?: string | null;
  socialHistory?: string | null;
  summary?: string | null;
  generatedAt: string;
}

export interface DoctorNote {
  id: number;
  caseId: number;
  doctorName: string;
  notes: string;
  createdAt: string;
}

export interface CaseFull {
  case: Case;
  session?: InterviewSession | null;
  report?: Report | null;
  doctorNotes: DoctorNote[];
}

export interface DoctorReview {
  case: Case;
  patient: Patient;
  profile?: PatientProfile | null;
  session?: InterviewSession | null;
  report?: Report | null;
  doctorNotes: DoctorNote[];
}

export interface CreatePatientRequest {
  name: string;
  age?: number | null;
  gender?: string | null;
  occupation?: string | null;
  weight?: string | null;
  height?: string | null;
  maritalStatus?: string | null;
  email?: string | null;
  phone?: string | null;
}

export interface UpdatePatientRequest {
  name?: string | null;
  age?: number | null;
  gender?: string | null;
  occupation?: string | null;
  weight?: string | null;
  height?: string | null;
  maritalStatus?: string | null;
  email?: string | null;
  phone?: string | null;
}

export type PatientProfileRequestCurrentMedicationsItem = {
  [key: string]: unknown;
};

export type PatientProfileRequestAllergiesItem = { [key: string]: unknown };

export type PatientProfileRequestFamilyHistory = { [key: string]: unknown };

export interface PatientProfileRequest {
  chronicConditions?: string[];
  surgicalHistory?: string[];
  currentMedications?: PatientProfileRequestCurrentMedicationsItem[];
  allergies?: PatientProfileRequestAllergiesItem[];
  familyHistory?: PatientProfileRequestFamilyHistory;
  smokingStatus?: string | null;
  alcoholUse?: string | null;
}

export interface CreateCaseRequest {
  patientId: number;
}

/**
 * Map of system name to array of selected symptoms
 */
export type RosRequestSymptoms = { [key: string]: unknown };

export interface RosRequest {
  /** Map of system name to array of selected symptoms */
  symptoms: RosRequestSymptoms;
}

export type ChiefComplaintsRequestComplaintsItem = {
  symptom: string;
  duration: string;
};

export interface ChiefComplaintsRequest {
  complaints: ChiefComplaintsRequestComplaintsItem[];
}

export interface InterviewMessageRequest {
  message: string;
}

export interface DoctorNoteRequest {
  doctorName: string;
  notes: string;
}