# lib/api-zod/src/generated/types/case.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hakim Medical History Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { CaseChiefComplaintsItem } from "./caseChiefComplaintsItem";
import type { CaseRosSymptoms } from "./caseRosSymptoms";

export interface Case {
  id: number;
  patientId: number;
  status: string;
  rosSymptoms?: CaseRosSymptoms;
  chiefComplaints?: CaseChiefComplaintsItem[] | null;
  createdAt: Date;
  updatedAt: Date;
}