# lib/api-zod/src/generated/types/caseFull.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hakim Medical History Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { Case } from "./case";
import type { DoctorNote } from "./doctorNote";
import type { InterviewSession } from "./interviewSession";
import type { Report } from "./report";

export interface CaseFull {
  case: Case;
  session?: InterviewSession | null;
  report?: Report | null;
  doctorNotes: DoctorNote[];
}