# lib/api-zod/src/generated/types/chiefComplaintsRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hakim Medical History Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { ChiefComplaintsRequestComplaintsItem } from "./chiefComplaintsRequestComplaintsItem";

export interface ChiefComplaintsRequest {
  complaints: ChiefComplaintsRequestComplaintsItem[];
}