# lib/api-zod/src/generated/types/doctorNote.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hakim Medical History Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface DoctorNote {
  id: number;
  caseId: number;
  doctorName: string;
  notes: string;
  createdAt: Date;
}