# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hakim Medical History Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface Patient {
  id: number;
  name: string;
  age?: number | null;
  gender?: string | null;
  occupation?: string | null;
  weight?: string | null;
  height?: string | null;
  maritalStatus?: string | null;
  email?: string | null;
  phone?: string | null;
  createdAt: Date;
  updatedAt: Date;
}