# lib/api-zod/src/generated/types/patientProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hakim Medical History Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { PatientProfileAllergiesItem } from "./patientProfileAllergiesItem";
import type { PatientProfileCurrentMedicationsItem } from "./patientProfileCurrentMedicationsItem";
import type { PatientProfileFamilyHistory } from "./patientProfileFamilyHistory";

export interface PatientProfile {
  id: number;
  patientId: number;
  chronicConditions?: string[];
  surgicalHistory?: string[];
  currentMedications?: PatientProfileCurrentMedicationsItem[];
  allergies?: PatientProfileAllergiesItem[];
  familyHistory?: PatientProfileFamilyHistory;
  smokingStatus?: string | null;
  alcoholUse?: string | null;
  createdAt: Date;
}