# lib/api-zod/src/generated/types/rosRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hakim Medical History Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { RosRequestSymptoms } from "./rosRequestSymptoms";

export interface RosRequest {
  /** Map of system name to array of selected symptoms */
  symptoms: RosRequestSymptoms;
}